+++ app/Http/Livewire/Shared/SelectUser.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Http\Livewire\Shared;
+
+use Livewire\Component;
+
+class SelectUser extends Component
+{
+    public function render()
+    {
+        return view('livewire.shared.select-user');
+    }
+}