+++ resources/views/course/show.blade.php
@@ -35,13 +35,13 @@
             @if ($course->video1)
             <main class="grid grid-cols-1 sm:grid-cols-5 gap-10 my-3 mx-3 md:mx-6 lg:mx-8">
                 <div class="col-span-5 sm:col-span-2 space-y-6 ">
-                    <div>
+                    <div class="pt-3">
                         @include('course.show.media')
                     </div>
                 </div>
                 <div class="col-span-5 sm:col-span-3 space-y-6">
                     @include('course.show.details')
-                    <div class="my-10">&nbsp;</div>
+                    <div class="my-20">&nbsp;</div>
                 </div>
             </main>
             @else
@@ -52,6 +52,7 @@
                 </div>
             </div>
             @endif
+            <div class="my-10">&nbsp;</div>
         </div>
         <aside class="w-full md:w-1/4">
             <div class="sticky top-6 space-y-4 px-3 sm:px-4 md:px-6 lg:px-8">